# Conflicts:
#	larsim/DetSim/SimWire_module.cc
#	larsim/DetSim/WienerFilterAna_module.cc
#	larsim/ElectronDrift/DriftElectronstoPlane_module.cc
#	larsim/ElectronDrift/ISCalculationSeparate.cc
#	larsim/ElectronDrift/ISCalculationSeparate.h
#	larsim/ElectronDrift/ShiftEdepSCE_module.cc
#	larsim/ElectronDrift/SimDriftElectrons_module.cc
#	larsim/EventGenerator/LightSource_module.cc
#	larsim/EventGenerator/POTaccumulator_module.cc
#	larsim/IonizationScintillation/ISCalc.cxx
#	larsim/IonizationScintillation/ISCalc.h
#	larsim/IonizationScintillation/ISCalcAna_module.cc
#	larsim/IonizationScintillation/ISCalcCorrelated.cxx
#	larsim/IonizationScintillation/ISCalcCorrelated.h
#	larsim/IonizationScintillation/ISCalcNESTLAr.cxx
#	larsim/IonizationScintillation/ISCalcNESTLAr.h
#	larsim/IonizationScintillation/ISCalcSeparate.cxx
#	larsim/IonizationScintillation/ISCalcSeparate.h
#	larsim/IonizationScintillation/IonAndScint_module.cc
#	larsim/LegacyLArG4/ConfigurablePhysicsList.hh
#	larsim/LegacyLArG4/ConfigurablePhysicsList.icc
#	larsim/LegacyLArG4/CustomPhysicsBuiltIns.cxx
#	larsim/LegacyLArG4/CustomPhysicsBuiltIns.hh
#	larsim/LegacyLArG4/CustomPhysicsFactory.hh
#	larsim/LegacyLArG4/CustomPhysicsTable.cxx
#	larsim/LegacyLArG4/CustomPhysicsTable.hh
#	larsim/LegacyLArG4/FastOpticalPhysics.cxx
#	larsim/LegacyLArG4/FastOpticalPhysics.h
#	larsim/LegacyLArG4/ISCalculation.cxx
#	larsim/LegacyLArG4/ISCalculation.h
#	larsim/LegacyLArG4/ISCalculationCorrelated.cxx
#	larsim/LegacyLArG4/ISCalculationCorrelated.h
#	larsim/LegacyLArG4/ISCalculationNEST.cxx
#	larsim/LegacyLArG4/ISCalculationNEST.h
#	larsim/LegacyLArG4/ISCalculationSeparate.cxx
#	larsim/LegacyLArG4/ISCalculationSeparate.h
#	larsim/LegacyLArG4/IonizationAndScintillation.cxx
#	larsim/LegacyLArG4/IonizationAndScintillation.h
#	larsim/LegacyLArG4/LArG4_module.cc
#	larsim/LegacyLArG4/LArSimChannelAna_module.cc
#	larsim/LegacyLArG4/LArVoxelReadout.cxx
#	larsim/LegacyLArG4/LArVoxelReadout.h
#	larsim/LegacyLArG4/LArVoxelReadoutGeometry.cxx
#	larsim/LegacyLArG4/LArVoxelReadoutGeometry.h
#	larsim/LegacyLArG4/MaterialPropertyLoader.cxx
#	larsim/LegacyLArG4/MaterialPropertyLoader.h
#	larsim/LegacyLArG4/NeutronHPphysics.cc
#	larsim/LegacyLArG4/NeutronHPphysics.hh
#	larsim/LegacyLArG4/OpBoundaryProcessSimple.cxx
#	larsim/LegacyLArG4/OpBoundaryProcessSimple.hh
#	larsim/LegacyLArG4/OpDetReadoutGeometry.cxx
#	larsim/LegacyLArG4/OpDetReadoutGeometry.h
#	larsim/LegacyLArG4/OpDetSensitiveDetector.cxx
#	larsim/LegacyLArG4/OpDetSensitiveDetector.h
#	larsim/LegacyLArG4/OpFastScintillation.cxx
#	larsim/LegacyLArG4/OpFastScintillation.hh
#	larsim/LegacyLArG4/OpticalPhysics.cxx
#	larsim/LegacyLArG4/OpticalPhysics.hh
#	larsim/LegacyLArG4/ParticleListAction.cxx
#	larsim/LegacyLArG4/ParticleListAction.h
#	larsim/LegacyLArG4/PhysicsList.cxx
#	larsim/MCCheater/BackTracker.cc
#	larsim/MCCheater/BackTracker.h
#	larsim/MCCheater/BackTracker.tcc
#	larsim/MCCheater/BackTrackerService.h
#	larsim/MCCheater/BackTrackerService_service.cc
#	larsim/MCCheater/CheckBackTracking_module.cc
#	larsim/MCCheater/RecoCheckAna_module.cc
#	larsim/PhotonPropagation/PDFastSimPAR_module.cc
#	larsim/PhotonPropagation/PhotonLibrary.cxx
#	larsim/PhotonPropagation/PhotonLibrary.h
#	larsim/PhotonPropagation/PhotonLibraryPropagation_module.cc
#	larsim/PhotonPropagation/PhotonVisibilityService_service.cc
#	larsim/SimFilters/FilterSimPhotonTime_module.cc
#	larsim/Simulation/PhotonVoxels.cxx
#	larsim/Simulation/PhotonVoxels.h
#	larsim/Simulation/SimListUtils.cxx
#	larsim/Simulation/SimListUtils.h
#	ups/product_deps

+++ larsim/PhotonPropagation/PDFastSimPAR_module.cc
@@ -404,11 +404,11 @@ namespace phot {
 
     void Initialization();
 
-    int VUVHits(int Nphotons_created,
+    int VUVHits(double Nphotons_created,
                 TVector3 ScintPoint,
                 TVector3 OpDetPoint,
                 int optical_detector_type);
-    int VISHits(int Nphotons_created,
+    int VISHits(double Nphotons_created,
                 TVector3 ScintPoint,
                 TVector3 OpDetPoint,
                 int optical_detector_type);
@@ -756,7 +756,7 @@ namespace phot {
 
   //......................................................................
   int
-  PDFastSimPAR::VUVHits(int Nphotons_created,
+  PDFastSimPAR::VUVHits(double Nphotons_created,
                         TVector3 ScintPoint,
                         TVector3 OpDetPoint,
                         int optical_detector_type)
@@ -839,7 +839,7 @@ namespace phot {
 
   //......................................................................
   int
-  PDFastSimPAR::VISHits(int Nphotons_created,
+  PDFastSimPAR::VISHits(double Nphotons_created,
                         TVector3 ScintPoint,
                         TVector3 OpDetPoint,
                         int optical_detector_type)
